requirements.txt
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.9
pandas==2.2.2
sqlalchemy==2.0.32
tenacity==9.0.0
pyyaml==6.0.2
